Refactor(backend): separate route, controller, and request base roles and functions
