+++ app/Console/Kernel.php
@@ -22,6 +22,6 @@ class Kernel extends ConsoleKernel
     {
         $this->load(__DIR__.'/Commands');
 
-        require base_path('routes/console.php');
+        require base_path('routes/console/console.php');
     }
 }
