+++ app/Http/Controllers/Kader/ArtikelResource.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Kader;
 
+use App\Http\Controllers\Controller;
 use App\Models\Artikel;
 use Illuminate\Http\Request;
