+++ app/Http/Controllers/Kader/InformasiController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Kader;
 
+use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 
 class InformasiController extends Controller
