+++ app/Http/Controllers/Kader/LansiaResource.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Kader;
 
+use App\Http\Controllers\Controller;
+use App\Http\Controllers\Kegiatan;
 use App\Models\Pemeriksaan;
-use App\Models\PemeriksaanLansia;
-use App\Models\Penduduk;
 use Carbon\Carbon;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
