+++ app/Http/Controllers/Ketua/BantuanController.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Ketua;
 
-use Illuminate\Http\Request;
+use App\Http\Controllers\Controller;
 
 class BantuanController extends Controller
 {
