+++ app/Http/Controllers/Ketua/PenerimaController.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Ketua;
 
-use Illuminate\Http\Request;
+use App\Http\Controllers\Controller;
 
 class PenerimaController extends Controller
 {
