+++ app/Http/Controllers/Shared/DataTablesController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Shared;
 
+use App\Http\Controllers\Controller;
 use Exception;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
@@ -70,6 +71,6 @@ class DataTablesController extends Controller
             })
             ->rawColumns(['aksi'])
             ->make();
-    }    
-    
+    }
+
 }
