+++ app/Http/Controllers/Shared/ProfileController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Shared;
 
+use App\Http\Controllers\Controller;
 use App\Models\Penduduk;
 use App\Models\User;
 use Illuminate\Http\RedirectResponse;
