+++ app/Http/Requests/Kader/Bayi/StoreBayiRequest.php
@@ -1,14 +1,11 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Kader\Bayi;
 
-use Carbon\Carbon;
 use Illuminate\Foundation\Http\FormRequest;
+use Illuminate\Validation\Rule;
 
-/**
- *
- */
-class StoreKegiatanRequest extends FormRequest
+class StoreBayiRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
@@ -25,17 +22,15 @@ class StoreKegiatanRequest extends FormRequest
      */
     protected function prepareForValidation(): void
     {
-        $this->merge([
-            'kader_id' => auth()->user()->kaders[0]->kader_id,
-            'tgl_kegiatan' => Carbon::create($this->input('year'), $this->input('month'), $this->input('day'))->format('Y-m-d')
-        ]);
-        $this->request->replace($this->only([
-            'kader_id',
-            'nama',
-            'tgl_kegiatan',
-            'jam_mulai',
-            'tempat'
-        ]));
+        $this->request->replace($this
+            ->only([
+                'lingkar_kepala',
+                'lingkar_lengan',
+                'asi',
+                'kenaikan',
+                'data_kb'
+            ])
+        );
     }
 
     /**
@@ -46,34 +41,32 @@ class StoreKegiatanRequest extends FormRequest
     public function rules(): array
     {
         return [
-            'kader_id' => [
+            'lingkar_kepala' => [
                 'bail',
                 'required',
-                'exists:kaders',
+                'numeric'
             ],
-            'nama' => [
+            'lingkar_lengan' => [
+                'bail',
+                'required',
+                'numeric'
+            ],
+            'asi' => [
+                'bail',
+                'required',
+                Rule::in(['iya', 'tidak'])
+            ],
+            'kenaikan' => [
+                'bail',
+                'required',
+                Rule::in(['iya', 'tidak'])
+            ],
+            'data_kb' => [
                 'bail',
                 'required',
                 'string',
-                'max:100',
-                'min:5'
+                'max:50'
             ],
-            'tgl_kegiatan' => [
-                'bail',
-                'required',
-                'date'
-            ],
-            'jam_mulai' => [
-                'bail',
-                'required',
-            ],
-            'tempat' => [
-                'bail',
-                'required',
-                'string',
-                'max:200',
-                'min:5'
-            ]
         ];
     }
 }
