+++ app/Providers/BroadcastServiceProvider.php
@@ -14,6 +14,6 @@ class BroadcastServiceProvider extends ServiceProvider
     {
         Broadcast::routes();
 
-        require base_path('routes/channels.php');
+        require base_path('routes/channels/channels.php');
     }
 }
